requirements.txt
attrs==24.3.0
allure-pytest==2.13.5
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
chromedriver-autoinstaller==0.6.4
chromedriver-binary==133.0.6942.0.0
click==8.1.7
configobj==5.0.9
dynaconf==3.2.4
exceptiongroup==1.2.2
Faker==33.1.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.1
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
pony==0.7.19
psycopg2-binary==2.9.10
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
tomli==2.0.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0